refactor(pack): refactor domain activation
