+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/domain/activation/activation.routing.js
@@ -1,0 +1,11 @@
+import controller from './activation.controller';
+import template from './activation.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.domain-activation', {
+    url: '/domain/activation',
+    template,
+    controller,
+    controllerAs: 'PackDomainActivation',
+  });
+};
