+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/domain/activation/index.js
@@ -1,0 +1,23 @@
+import angular from 'angular';
+import ngOvhTelecomUniverseComponents from '@ovh-ux/ng-ovh-telecom-universe-components';
+import '@uirouter/angularjs';
+import '@ovh-ux/ng-translate-async-loader';
+import 'angular-translate';
+import 'ovh-api-services';
+
+import routing from './activation.routing';
+
+const moduleName = 'ovhManagerTelecomTelephonyBillingAccountManageContacts';
+
+angular
+  .module(moduleName, [
+    ngOvhTelecomUniverseComponents,
+    'ngTranslateAsyncLoader',
+    'pascalprecht.translate',
+    'ovh-api-services',
+    'ui.router',
+  ])
+  .config(routing)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;
